+++ packages/wallets/wc/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @swapkit/wallet-wc
 
+## 1.7.0
+
+### Minor Changes
+
+- [#1185](https://github.com/thorswap/SwapKit/pull/1185) [`6b9a5d7`](https://github.com/thorswap/SwapKit/commit/6b9a5d7dc2dad7d74052ce8308d91bf1da8eb951) Thanks [@towanTG](https://github.com/towanTG)! - Adds alchemy api and refactors wallets api handling internally
+
+### Patch Changes
+
+- Updated dependencies [[`6b9a5d7`](https://github.com/thorswap/SwapKit/commit/6b9a5d7dc2dad7d74052ce8308d91bf1da8eb951)]:
+  - @swapkit/toolbox-cosmos@1.9.0
+  - @swapkit/helpers@2.5.0
+  - @swapkit/toolbox-evm@1.9.0
+
 ## 1.6.11
 
 ### Patch Changes